requirements.txt
streamlit>=1.32.0
openai>=1.12.0
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
crawl4ai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
langgraph>=0.2.0
typing-extensions>=4.9.0
apscheduler>=3.10.0
